Card Refactoring
