+++ Assets/Scripts/Card/Card/Card.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c636b79d6948ccc46b1152acae5ec916
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
