+++ Assets/Scripts/Card/Card/CardData.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0085985a56a413c4395706471ac93330
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
