+++ Assets/Scripts/Card/Card/CardData/ArtifactCardData.cs
@@ -1,0 +1,7 @@
+using UnityEngine;
+
+[CreateAssetMenu(fileName = "Artifact Card", menuName = "Card/Artifact Card")]
+public class ArtifactCardData : CardData
+{
+    public CardEffect[] effects;
+}
