+++ Assets/Scripts/Card/Card/CardData/ArtifactCardData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 171d6ce6218740f43957c6dd0c3ac486
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
