+++ Assets/Scripts/Card/Card/CardData/BattleCardData.cs
@@ -1,0 +1,8 @@
+using UnityEngine;
+
+[CreateAssetMenu(fileName = "Battle Card", menuName = "Card/Battle Card")]
+public class BattleCardData : CardData
+{
+    public int cardCost;
+    public CardEffect[] effects;
+}
