+++ Assets/Scripts/Card/Card/CardData/BattleCardData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1d0f5d2874c862d439f99e94b63cce13
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
