+++ Assets/Scripts/Card/Card/CardData/CardData.cs
@@ -1,0 +1,8 @@
+using UnityEngine;
+
+public abstract class CardData : ScriptableObject
+{
+    public string cardName;
+    public Sprite cardImage;
+    public string cardDescription;
+}
