+++ Assets/Scripts/Card/Card/CardData/ExplorationCardData.cs
@@ -1,0 +1,8 @@
+using UnityEngine;
+
+[CreateAssetMenu(fileName = "Exploration Card", menuName = "Card/Exploration Card")]
+public class ExplorationCardData : CardData
+{
+    public int requiredSteps;
+    public CardEffect[] effects;
+}
