+++ Assets/Scripts/Card/Card/CardData/ExplorationCardData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e4a93215fc8333e4793bd93c23d193d3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
