+++ Assets/Scripts/Card/Card/CardManager.cs
@@ -5,6 +5,5 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "CardDatas", menuName = "Card/Cards")]
 public class CardManager : ScriptableObject
 {
-    public Card[] cards1;
     public BattleCardData[] cards;
 }
